Create the Booker class
Add the Booker class. The Booker class manages the stages of booking
a cinema seat. Select an array of seats for a projection and store
them temporarily.

Display a grid of the projection seating allocations. As the Booker
adds new temporary seat allocations, add these to a temporary grid
showing the remaining available seats for the projection. When a
Booker switches to a different projection, reset the seatReservationList
so that there is no conflict between seat reservations for different
projections.