Remove Booker seat reservation
Enable Booker to remove seat reservations from its list. This allows
users to deselect seats from the projection seat grid, without needing
to reset the entire booking process.

Booker now takes a Projection object in its constructor. When the seats
are being selected, or for any other method in the Booker class, the
projection will always be the same. This change is made to remove the
need for a resetProjection() method whenever a different projection is
selected. A Booking can now not be made for multiple projections.