Add Ticket Reschedule feature
Create Transferer class. Create abstract TicketManager class. Set
Booker as inheriting from TicketManager class. Create abstract
SeatAllocation class. Rename SeatReservation class to Reservation, and
allow Reservation and Ticket to inherit from SeatAllocation. These
name changes and abstract classes group similar classes together.

In the Transferer class, accept a Projection object and a Ticket object
Review the current seating grid for the Projection, select a seat to
transfer the ticket to. When finalizeCashPayment is called (or
finalizeCardPayment) the ticket details are updated, and a new payment
is added to the Booking if the new ticket is more expensive than the
old ticket. If the ticket is the same or cheaper, then no new payment
is created, and the ticket-details are simply updated.