Refactor the long CLI class
Break the CLI class into smaller mothods.

Refactor the code so that each page has a method with a return statement
as an int. This is returned to the run() method. If the return value is
6, then quit the application. Otherwise, display the requested page.
Each page has a pageId from 1 to 19. displayShowObject methods do not have
pageIds because they are treated differently. They accept an Object as a
parameter, so they must be treated differently. When a displayShowObject
method is called, pass in an Object, return an int and bounce this int to
the run() method using another return statement.