Add Review and Rating to Cinema reviews
Add the review and rating methods to CLI. The user enters the
TicketId that is printed on their ticket. If the show has not yet
happened, a review cannot be logged. If a review has already been
logged for that ticket, the review cannot be overwritten. One
review per ticket.

Add a confirmation question to the Application Quit process, so
that the user confirms that they want to quit before quitting.