Add Gui structure with FilmsIndex
Add the Gui page template. List Films in the CENTER. When a film is
clicked, display the film details in the EAST.

The MouseListener uses Object so that it can be reused for the other
object types. The MouseListener passes an Object to the updateEastPanel
method. In the updateEastPanel method, an instanceof check tests to see
the object type and updates the eastPanel with the details about the
object.