Create addShow page with FilmList and ScreenList
Create the addShow page. User selects a Film from the list. User
selects a Screen from the list. User enters the date of the show,
and the priceRegular and priceVip. The show is saved and the user
is redirected to the ShowsIndex page where they can see the new
show.

The function uses a formData HashMap that is stored as an instance
variable.