Add Ticket Transfer functionality
Add Ticket Transfer feature. Add Move button to Bookings menu. Update
getSeatingGridButtons to take a ticketManager param. The param lets the
getSeatingGridButtons method to be used by both the Booker and the
Transferer.

NoChargeTransfers are processed with a single click. If there is a transfer
surcharge transferer.getSurcharge>0)
then display a dialog box to select card or cash payment. If [CARD] is
selected, display a second dialog box requesting the card-payment
reference number.