Add date validation for transfer, booking, review and shows
Add validation to ensure that transfers can only be made with tickets
for shows in the future, and they can only be moved to shows in the
future. Add validation to ensure that bookings can only be moved to shows
*made for shows in the future. Add validation so that reviews can
only be made for tickets in the past.

BugFix: Remove ActionListener for selected seat if the the seat
is unavailable.

BugFix: Ignore ticket in seatingGridButtons if transferring ticket.
Do not ignore ticket if the new show is different to the ticket's
previous show.