Move getNextId() methods to their own classes
Move the getNextId() methods from the Cinema class and move
each to the class of the object that it relates to. Store
the nextUnusedId as a private static variable in those classes
and create a public static getNextId() method to get the
next id and increment.

This reduces the number of methods in the Cinema class and
increases the level of cohesion in the classes.